make city static & refactor
