+++ index.html
@@ -182,7 +182,6 @@
               id="modify-street"
               placeholder="Hauptstraße 1"
               value="Hauptstraße 1"
-              required
               disabled
             />
           </div>
@@ -193,7 +192,6 @@
               id="modify-zipcode"
               placeholder="12489"
               value="12489"
-              required
               disabled
             />
           </div>
@@ -204,7 +202,6 @@
               id="modify-city"
               placeholder="Berlin"
               value="Berlin"
-              required
               disabled
             />
           </div>
