+++ js/routingService.js
@@ -1,4 +1,4 @@
-import { deactivateById, reactivateById } from './domHelper.js';
+import { deactivateById, reactivateById, setValueById } from './domHelper.js';
 
 const navigateToScreenById = (selectedScreen, isAdmin) => {
   const screens = [
@@ -20,4 +20,23 @@ const navigateOnClick = (id, destination) => {
     navigateToScreenById(destination);
 };
 
-export { navigateToScreenById, navigateOnClick };
+const navigateToPrefilledDetails = (element, location) => {
+  element.onclick = (e) => {
+    e?.preventDefault();
+    const { id, title, description, street, zipcode, lat, lon, score } =
+      location;
+    setValueById('modify-title', title);
+    setValueById('modify-description', description);
+    setValueById('modify-street', street);
+    setValueById('modify-zipcode', zipcode);
+    setValueById('modify-city', 'Berlin');
+    setValueById('modify-latitude', lat);
+    setValueById('modify-longitude', lon);
+    setValueById('modify-pollution', score);
+    document.getElementById('hidden-id-field').innerText = id;
+
+    navigateToScreenById('update-screen');
+  };
+};
+
+export { navigateToScreenById, navigateOnClick, navigateToPrefilledDetails };
